Added property url
