+++ src/main/java/property/Property.java
@@ -1,0 +1,5 @@
+package property;
+
+public class Property {
+    public static String url = System.getProperty("url", "https://stellarburgers.nomoreparties.site");
+}
