+++ src/main/java/stellarburgers/User.java
@@ -1,4 +1,4 @@
-package stellarBurgers;
+package stellarburgers;
 
 public class User {
     private String email;
